+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 # cordova-sqlite-evmax-build-free 0.2.0-dev
 
+- add ICU support for Android & iOS using built-in platform ICU on each platform (requires Android 12+)
+
 ## cordova-sqlite-evmax-feat-android-db-location 0.0.3-dev
 
 ## cordova-sqlite-evmax-common-free 0.0.8-dev